+++ src/AggregateFunction.ts
@@ -5,6 +5,8 @@ import { BuilderData } from './builder'
 import { ComparisonOperator } from './operators'
 import { Binder } from './binder'
 import { IStatementGiver } from './models/IStatementGiver'
+import { Column } from './columns'
+import { ItemInfo } from './ItemInfo'
 
 export enum AggregateFunctionEnum {
   SUM = 'SUM',
@@ -14,13 +16,14 @@ export enum AggregateFunctionEnum {
   MIN = 'MIN',
 }
 
-export class AggregateFunction implements IStatementGiver{
-  constructor(private readonly funcName: AggregateFunctionEnum, private readonly expression: Expression) {
+export class AggregateFunction implements IStatementGiver {
+  private readonly unique: symbol = Symbol()
+  constructor(public readonly funcName: AggregateFunctionEnum, private readonly expression: Expression) {
     if (expression.type !== ExpressionType.NUMBER)
       throw new Error('Expression Type must be number in aggregate function')
   }
 
-  public as(alias: string): SelectItemInfo {
+  public as(alias: string): ItemInfo {
     return new SelectItemInfo(this, alias)
   }
 
@@ -36,7 +39,7 @@ export class AggregateFunction implements IStatementGiver{
     return new Condition(new Expression(this), ComparisonOperator.NotEqual, new Expression(value))
   }
 
-  public ne$(value:number): Condition {
+  public ne$(value: number): Condition {
     return new Condition(new Expression(this), ComparisonOperator.NotEqual, new Expression(new Binder(value)))
   }
 
@@ -76,5 +79,9 @@ export class AggregateFunction implements IStatementGiver{
     if (this.expression.rightOperand === undefined || this.expression.rightOperand.type === ExpressionType.NOT_EXIST)
       return `${this.funcName}(${this.expression.getStmt(data)})`
     return `${this.funcName}${this.expression.getStmt(data)}`
+  }
+
+  public getColumns(): Column[] {
+    return this.expression.getColumns()
   }
 }