+++ src/steps/DeleteFromStep.ts
@@ -4,6 +4,11 @@ import { Condition } from '../models/Condition'
 import { DeleteWhereStep } from './DeleteWhereStep'
 import { LogicalOperator } from '../operators'
 import { MoreThanOneWhereStepError } from '../errors'
+import { ReturningStep } from './stepInterfaces'
+import { returnStepOrThrow } from '../util'
+import { PrimitiveType } from '../models/types'
+import { ReturningItem } from '../ReturningItemInfo'
+import { ItemInfo } from '../ItemInfo'
 
 export class DeleteFromStep extends BaseStep {
   constructor(protected data: BuilderData) { super(data) }
@@ -17,5 +22,9 @@ export class DeleteFromStep extends BaseStep {
     }
     this.addWhereParts(cond1, op1, cond2, op2, cond3)
     return new DeleteWhereStep(this.data)
+  }
+
+  public returning(...items: (ItemInfo|ReturningItem|PrimitiveType)[]): ReturningStep {
+    return returnStepOrThrow(this.data.step).returning(...items)
   }
 }